BONUS: Updated functions to use decimal instead of python's default integer division, so now invoices get proper dollar amounts. Updated make_invoices to be way simpler.

+++ accounting/models.py
@@ -14,7 +14,7 @@ class Policy(db.Model):
     effective_date = db.Column(u'effective_date', db.DATE(), nullable=False)
     status = db.Column(u'status', db.Enum(u'Active', u'Canceled', u'Expired'), default=u'Active', nullable=False)
     billing_schedule = db.Column(u'billing_schedule', db.Enum(u'Annual', u'Two-Pay', u'Quarterly', u'Monthly'), default=u'Annual', nullable=False)
-    annual_premium = db.Column(u'annual_premium', db.INTEGER(), nullable=False)
+    annual_premium = db.Column(u'annual_premium', db.NUMERIC(), nullable=False)
     named_insured = db.Column(u'named_insured', db.INTEGER(), db.ForeignKey('contacts.id'))
     agent = db.Column(u'agent', db.INTEGER(), db.ForeignKey('contacts.id'))
     canceled_date = db.Column(u'canceled_date', db.DATE())
@@ -54,7 +54,7 @@ class Invoice(db.Model):
     bill_date = db.Column(u'bill_date', db.DATE(), nullable=False)
     due_date = db.Column(u'due_date', db.DATE(), nullable=False)
     cancel_date = db.Column(u'cancel_date', db.DATE(), nullable=False)
-    amount_due = db.Column(u'amount_due', db.INTEGER(), nullable=False)
+    amount_due = db.Column(u'amount_due', db.NUMERIC(), nullable=False)
     deleted = db.Column(u'deleted', db.Boolean, default=False, server_default='0', nullable=False)
 
     def __init__(self, policy_id, bill_date, due_date, cancel_date, amount_due):
@@ -74,7 +74,7 @@ class Payment(db.Model):
     id = db.Column(u'id', db.INTEGER(), primary_key=True, nullable=False)
     policy_id = db.Column(u'policy_id', db.INTEGER(), db.ForeignKey('policies.id'), nullable=False)
     contact_id = db.Column(u'contact_id', db.INTEGER(), db.ForeignKey('contacts.id'), nullable=False)
-    amount_paid = db.Column(u'amount_paid', db.INTEGER(), nullable=False)
+    amount_paid = db.Column(u'amount_paid', db.NUMERIC(), nullable=False)
     transaction_date = db.Column(u'transaction_date', db.DATE(), nullable=False)
 
     def __init__(self, policy_id, contact_id, amount_paid, transaction_date):
